Move rotation strategy classes into subpackage
Also move rotation strategy bindings beneath the actual binding classes.
