+++ graylog2-server/src/main/java/org/graylog2/indexer/rotation/RotationStrategyBindings.java
@@ -17,19 +17,17 @@
 
 package org.graylog2.indexer.rotation;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.auto.value.AutoValue;
+import org.graylog2.indexer.rotation.strategies.MessageCountRotationStrategy;
+import org.graylog2.indexer.rotation.strategies.SizeBasedRotationStrategy;
+import org.graylog2.indexer.rotation.strategies.TimeBasedRotationStrategy;
+import org.graylog2.plugin.PluginModule;
 
-@JsonAutoDetect
-@AutoValue
-public abstract class MessageCountRotationStrategyConfig {
-    @JsonProperty("max_docs_per_index")
-    public abstract int maxDocsPerIndex();
-
-    @JsonCreator
-    public static MessageCountRotationStrategyConfig create(@JsonProperty("max_docs_per_index") int maxDocsPerIndex) {
-        return new AutoValue_MessageCountRotationStrategyConfig(maxDocsPerIndex);
+public class RotationStrategyBindings extends PluginModule {
+    @Override
+    protected void configure() {
+        addRotationStrategy(MessageCountRotationStrategy.class);
+        addRotationStrategy(SizeBasedRotationStrategy.class);
+        addRotationStrategy(TimeBasedRotationStrategy.class);
     }
+
 }
