+++ graylog2-server/src/main/java/org/graylog2/indexer/rotation/strategies/MessageCountRotationStrategy.java
@@ -14,7 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.indexer.rotation;
+
+package org.graylog2.indexer.rotation.strategies;
 
 import org.graylog2.indexer.Deflector;
 import org.graylog2.indexer.IndexNotFoundException;
@@ -43,7 +44,7 @@ public class MessageCountRotationStrategy extends AbstractRotationStrategy {
 
     @Nullable
     @Override
-    protected AbstractRotationStrategy.Result shouldRotate(String index) {
+    protected Result shouldRotate(String index) {
         final MessageCountRotationStrategyConfig config = clusterConfigService.get(MessageCountRotationStrategyConfig.class);
 
         if (config == null) {
