+++ graylog2-server/src/test/java/org/graylog2/indexer/rotation/strategies/MessageCountRotationStrategyTest.java
@@ -14,13 +14,11 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.indexer.rotation;
 
-import org.elasticsearch.action.admin.indices.stats.CommonStats;
-import org.elasticsearch.cluster.routing.ShardRouting;
-import org.elasticsearch.index.store.StoreStats;
+package org.graylog2.indexer.rotation.strategies;
+
 import org.graylog2.indexer.Deflector;
-import org.graylog2.indexer.indices.IndexStatistics;
+import org.graylog2.indexer.IndexNotFoundException;
 import org.graylog2.indexer.indices.Indices;
 import org.graylog2.plugin.cluster.ClusterConfigService;
 import org.junit.Test;
@@ -28,8 +26,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import java.util.Collections;
-
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
@@ -37,7 +34,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SizeBasedRotationStrategyTest {
+public class MessageCountRotationStrategyTest {
     @Mock
     private ClusterConfigService clusterConfigService;
 
@@ -49,33 +46,24 @@ public class SizeBasedRotationStrategyTest {
 
     @Test
     public void testRotate() throws Exception {
-        final CommonStats commonStats = new CommonStats();
-        commonStats.store = new StoreStats(1000, 0);
-        final IndexStatistics stats = IndexStatistics.create("name", commonStats, commonStats, Collections.<ShardRouting>emptyList());
-
-        when(indices.getIndexStats("name")).thenReturn(stats);
+        when(indices.numberOfMessages("name")).thenReturn(10L);
         when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100L));
+        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
 
-        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
+        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, times(1)).cycle();
         reset(deflector);
     }
 
-
     @Test
     public void testDontRotate() throws Exception {
-        final CommonStats commonStats = new CommonStats();
-        commonStats.store = new StoreStats(1000, 0);
-        final IndexStatistics stats = IndexStatistics.create("name", commonStats, commonStats, Collections.<ShardRouting>emptyList());
-
-        when(indices.getIndexStats("name")).thenReturn(stats);
+        when(indices.numberOfMessages("name")).thenReturn(1L);
         when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100000L));
+        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
 
-        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
+        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, never()).cycle();
@@ -84,12 +72,12 @@ public class SizeBasedRotationStrategyTest {
 
 
     @Test
-    public void testRotateFailed() throws Exception {
-        when(indices.getIndexStats("name")).thenReturn(null);
+    public void testIndexUnavailable() throws Exception {
+        doThrow(IndexNotFoundException.class).when(indices).numberOfMessages("name");
         when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100));
+        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
 
-        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
+        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, never()).cycle();
