+++ graylog2-server/src/test/java/org/graylog2/indexer/rotation/strategies/SizeBasedRotationStrategyTest.java
@@ -14,10 +14,14 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.indexer.rotation;
 
+package org.graylog2.indexer.rotation.strategies;
+
+import org.elasticsearch.action.admin.indices.stats.CommonStats;
+import org.elasticsearch.cluster.routing.ShardRouting;
+import org.elasticsearch.index.store.StoreStats;
 import org.graylog2.indexer.Deflector;
-import org.graylog2.indexer.IndexNotFoundException;
+import org.graylog2.indexer.indices.IndexStatistics;
 import org.graylog2.indexer.indices.Indices;
 import org.graylog2.plugin.cluster.ClusterConfigService;
 import org.junit.Test;
@@ -25,7 +29,8 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import static org.mockito.Mockito.doThrow;
+import java.util.Collections;
+
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
@@ -33,7 +38,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class MessageCountRotationStrategyTest {
+public class SizeBasedRotationStrategyTest {
     @Mock
     private ClusterConfigService clusterConfigService;
 
@@ -45,24 +50,33 @@ public class MessageCountRotationStrategyTest {
 
     @Test
     public void testRotate() throws Exception {
-        when(indices.numberOfMessages("name")).thenReturn(10L);
-        when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
+        final CommonStats commonStats = new CommonStats();
+        commonStats.store = new StoreStats(1000, 0);
+        final IndexStatistics stats = IndexStatistics.create("name", commonStats, commonStats, Collections.<ShardRouting>emptyList());
 
-        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
+        when(indices.getIndexStats("name")).thenReturn(stats);
+        when(deflector.getNewestTargetName()).thenReturn("name");
+        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100L));
+
+        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, times(1)).cycle();
         reset(deflector);
     }
 
+
     @Test
     public void testDontRotate() throws Exception {
-        when(indices.numberOfMessages("name")).thenReturn(1L);
-        when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
+        final CommonStats commonStats = new CommonStats();
+        commonStats.store = new StoreStats(1000, 0);
+        final IndexStatistics stats = IndexStatistics.create("name", commonStats, commonStats, Collections.<ShardRouting>emptyList());
 
-        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
+        when(indices.getIndexStats("name")).thenReturn(stats);
+        when(deflector.getNewestTargetName()).thenReturn("name");
+        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100000L));
+
+        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, never()).cycle();
@@ -71,12 +85,12 @@ public class MessageCountRotationStrategyTest {
 
 
     @Test
-    public void testIndexUnavailable() throws Exception {
-        doThrow(IndexNotFoundException.class).when(indices).numberOfMessages("name");
+    public void testRotateFailed() throws Exception {
+        when(indices.getIndexStats("name")).thenReturn(null);
         when(deflector.getNewestTargetName()).thenReturn("name");
-        when(clusterConfigService.get(MessageCountRotationStrategyConfig.class)).thenReturn(MessageCountRotationStrategyConfig.create(5));
+        when(clusterConfigService.get(SizeBasedRotationStrategyConfig.class)).thenReturn(SizeBasedRotationStrategyConfig.create(100));
 
-        final MessageCountRotationStrategy strategy = new MessageCountRotationStrategy(indices, deflector, clusterConfigService);
+        final SizeBasedRotationStrategy strategy = new SizeBasedRotationStrategy(indices, deflector, clusterConfigService);
 
         strategy.rotate();
         verify(deflector, never()).cycle();
